refactor: import and component
