--- src/components/ui/BaseCard.tsx
@@ -1,1 +1,0 @@
-// Wrapper cho các card component
