--- src/components/ui/HistoryCard.tsx
@@ -1,1 +1,0 @@
-// Lịch sử interview card component
