--- src/components/ui/Modal.tsx
@@ -1,1 +1,0 @@
-// Modal chung có thể custom title, content
